+++ tests/grippeweb/test_extract.py
@@ -1,5 +1,3 @@
-from uuid import UUID
-
 import pytest
 
 from mex.common.ldap.models import LDAPFunctionalAccount, LDAPPerson
@@ -41,37 +39,27 @@ def test_extract_columns_by_table_and_column_name() -> None:
 @pytest.mark.usefixtures("mocked_ldap", "mocked_grippeweb")
 def test_extract_ldap_actors_for_functional_accounts(
     grippeweb_resource_mappings: list[ResourceMapping],
+    ldap_contact_point: LDAPFunctionalAccount,
 ) -> None:
     ldap_actors = extract_ldap_actors_for_functional_accounts(
         grippeweb_resource_mappings
     )
-    expected = LDAPFunctionalAccount(
-        objectGUID=UUID("00000000-0000-4000-8000-000000000004"),
-        sAMAccountName="ContactC",
-        mail=["email@email.de", "contactc@rki.de"],
-        ou=["Funktion"],
-    )
-    assert ldap_actors[0] == expected
+
+    assert ldap_actors[0] == ldap_contact_point
 
 
 @pytest.mark.usefixtures("mocked_ldap", "mocked_grippeweb")
 def test_extract_ldap_persons(
     grippeweb_resource_mappings: list[ResourceMapping],
     grippeweb_access_platform: AccessPlatformMapping,
+    ldap_roland_resolved: LDAPPerson,
 ) -> None:
     ldap_persons = extract_ldap_persons(
         grippeweb_resource_mappings, grippeweb_access_platform
     )
-    expected = LDAPPerson(
-        objectGUID=UUID(int=1, version=4),
-        mail=["test_person@email.de"],
-        department="PARENT-UNIT",
-        displayName="Resolved, Roland",
-        employeeID="42",
-        givenName=["Roland"],
-        sn="Resolved",
-    )
-    assert ldap_persons[0] == expected
+
+    assert len(ldap_persons) == 3
+    assert ldap_persons[0] == ldap_roland_resolved
 
 
 @pytest.mark.usefixtures("mocked_wikidata", "mocked_grippeweb")