+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- update dependencies mex-artificial >= 1.3, Faker >= 40
+
 ### Deprecated
 
 ### Removed