+++ CHANGELOG.md
@@ -10,6 +10,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - convenience function for LDAP lookup
+
+### Changes
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [1.8.1] - 2026-01-29
+
+### Added
+
+- enable testing against a mocked ldap container
+
+### Changes
+
+- update dependencies mex-artificial >= 1.3, Faker >= 40
+
+## [1.8.0] - 2026-01-28
+
+### Added
+
 - extraction of kvis sql tables
 - prepare testing against a mocked ldap container
 - primary source helper function: add caching
@@ -20,9 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/33ca3a
-
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/ad26e4
-
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/203934
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/c5ff3e
 - update mex-common dependency to 1.11
@@ -34,8 +57,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - datenkompass: publish activities by unit, filter and mapping update
 - switch package manager from pdm to uv
 
-### Deprecated
-
 ### Removed
 
 - tailored tests for blueant and endnote; now generalized in test_checks_main.py
@@ -44,8 +65,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - blueant: fix Person extraction and units written as MergedOrganization items
-
-### Security
 
 ## [1.7.0] - 2025-12-11
 
@@ -60,6 +79,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pin Mailpit Docker image version to v1.28
 
 ### Fixed
+
 - synopse: fix extraction of involved units
 
 ## [1.6.0] - 2025-11-28